Fix for python3 x86

+++ Shared_files/AtanTable.py
@@ -5,7 +5,7 @@ import sys
 from mpmath import mp
 mp.dps = 50
 
-N = long(sys.argv[1]) if len(sys.argv) > 1 else 500000
+N = int(sys.argv[1]) if len(sys.argv) > 1 else 500000
 
 dx = mp.mpf("0.002")
